chore: PathVariable 수정
